Clean(TM) Code(C)

+++ code/modules/mob/living/silicon/pai/pai.dm
@@ -1,3 +1,6 @@
+#define PAI_HUD_SEC (1<<0)
+#define PAI_HUD_MED (1<<0)
+
 /mob/living/silicon/pai
 	name = "pAI"
 	icon = 'icons/mob/pai.dmi'
@@ -10,63 +13,88 @@
 	health = 500
 	maxHealth = 500
 	layer = BELOW_MOB_LAYER
-	var/obj/item/instrument/piano_synth/internal_instrument
 	silicon_privileges = PRIVILEDGES_PAI
+	mobility_flags = NONE
 
-	var/network = "ss13"
-	var/obj/machinery/camera/current = null
-
-	var/ram = 100	// Used as currency to purchase different abilities
+	/// Used as currency to purchase different abilities
+	var/ram = 100
+	/// Installed software
 	var/list/software = list()
-	var/userDNA		// The DNA string of our assigned user
-	var/obj/item/paicard/card	// The card we inhabit
-	var/hacking = FALSE		//Are we hacking a door?
+	/// The card we inhabit
+	var/obj/item/paicard/card
+
+	/// User vars
+	
+	/// Name of the one who commands us
+	var/master
+	/// DNA string for owner verification
+	var/master_dna
+	/// The DNA string of our assigned user
+	var/userDNA
 
 	var/speakStatement = "states"
 	var/speakExclamation = "declares"
 	var/speakDoubleExclamation = "alarms"
 	var/speakQuery = "queries"
 
-	var/obj/item/radio/headset			// The pAI's headset
-	var/obj/item/pai_cable/cable		// The cable we produce and use when door or camera jacking
 
-	var/master				// Name of the one who commands us
-	var/master_dna			// DNA string for owner verification
+	/// HTML Ui vars
 
-// Various software-specific vars
+	/// General error reporting text contained here will typically be shown once and cleared
+	var/temp
+	/// Which screen our main window displays
+	var/screen
+	/// Which specific function of the main screen is being displayed
+	var/subscreen
 
-	var/temp				// General error reporting text contained here will typically be shown once and cleared
-	var/screen				// Which screen our main window displays
-	var/subscreen			// Which specific function of the main screen is being displayed
+	/// Various software-specific vars
+	
+	/// Are we hacking a door?
+	var/hacking = FALSE
+	/// Instrument for it's own synth
+	var/obj/item/instrument/piano_synth/internal_instrument
 
+	/// The pAI's headset
+	var/obj/item/radio/headset/headset
+	/// The cable we produce and use when door or camera jacking		
+	var/obj/item/pai_cable/cable
+
+	/// Cammera network access
+	var/network = "ss13"
+	var/obj/machinery/camera/current = null
+
+	/// The airlock being hacked
+	var/obj/machinery/door/hackdoor
+	/// 1-100, >= 100 means the hack is complete and will be reset upon next check
+	var/hackprogress = 0
+
+	/// Custom PDA
 	var/obj/item/pda/ai/pai/pda = null
 
-	var/secHUD = 0			// Toggles whether the Security HUD is active or not
-	var/medHUD = 0			// Toggles whether the Medical  HUD is active or not
+	/// Toggles whether the Security HUD is active or not
+	var/secHUD = FALSE
+	/// Toggles whether the Medical  HUD is active or not
+	var/medHUD = FALSE
 
-	var/datum/data/record/medicalActive1		// Datacore record declarations for record software
+	/// pAI hud flags (PAI_HUD_SEC|PAI_HUD_MED)
+	var/pAI_HUD_flags = NONE
+
+	/// Datacore record declarations for record software
+	var/datum/data/record/medicalActive1
 	var/datum/data/record/medicalActive2
 
-	var/datum/data/record/securityActive1		// Could probably just combine all these into one
+	/// Could probably just combine all these into one
+	var/datum/data/record/securityActive1
 	var/datum/data/record/securityActive2
-
-	var/obj/machinery/door/hackdoor		// The airlock being hacked
-	var/hackprogress = 0				// Possible values: 0 - 100, >= 100 means the hack is complete and will be reset upon next check
-
-	var/obj/item/integrated_signaler/signaler // AI's signaller
+	
+	/// pAI's signaller
+	var/obj/item/integrated_signaler/signaler 
 
 	var/encryptmod = FALSE
+	var/obj/item/card/id/access_card = null
+
 	var/holoform = FALSE
 	var/canholo = TRUE
-	var/obj/item/card/id/access_card = null
-	var/chassis = "repairbot"
-	var/dynamic_chassis
-	var/dynamic_chassis_sit = FALSE			//whether we're sitting instead of resting spritewise
-	var/dynamic_chassis_bellyup = FALSE		//whether we're lying down bellyup
-	var/list/possible_chassis			//initialized in initialize.
-	var/list/dynamic_chassis_icons		//ditto.
-	var/list/chassis_pixel_offsets_x	//stupid dogborgs
-
 	var/emitterhealth = 20
 	var/emittermaxhealth = 20
 	var/emitterregen = 0.25
@@ -75,86 +103,97 @@
 	var/emittercd = 50
 	var/emitteroverloadcd = 100
 
+	var/chassis = "repairbot"
+	var/dynamic_chassis
+	var/dynamic_chassis_sit = FALSE			//whether we're sitting instead of resting spritewise
+	var/dynamic_chassis_bellyup = FALSE		//whether we're lying down bellyup
+	var/list/possible_chassis			//initialized in initialize.
+	var/list/dynamic_chassis_icons		//ditto.
+	var/list/chassis_pixel_offsets_x	//stupid dogborgs
+
 	var/radio_short = FALSE
 	var/radio_short_cooldown = 3 MINUTES
 	var/radio_short_timerid
 
-	mobility_flags = NONE
 	var/silent = FALSE
 	var/brightness_power = 5
 
 	var/icon/custom_holoform_icon
 
 /mob/living/silicon/pai/Destroy()
-	QDEL_NULL(internal_instrument)
+	QDEL_NULL(radio)
+	QDEL_NULL(pda)
+	QDEL_NULL(signaler)
+	
+	if(internal_instrument) //installed, and does not exist on init
+		QDEL_NULL(internal_instrument)
+
 	if (loc != card)
 		card.forceMove(drop_location())
+
 	card.pai = null
 	card.cut_overlays()
 	card.add_overlay("pai-off")
 	GLOB.pai_list -= src
+
 	return ..()
 
 /mob/living/silicon/pai/Initialize()
 	var/obj/item/paicard/P = loc
-	START_PROCESSING(SSfastprocess, src)
-	GLOB.pai_list += src
-	make_laws()
 	if(!istype(P)) //when manually spawning a pai, we create a card to put it into.
 		var/newcardloc = P
 		P = new /obj/item/paicard(newcardloc)
 		P.setPersonality(src)
+
 	forceMove(P)
 	card = P
+
 	signaler = new(src)
-	if(!radio)
-		radio = new /obj/item/radio/headset/silicon/pai(src)
+	pda = new(src) //let it finish it's own init
+	radio = new /obj/item/radio/headset/silicon/pai(src) //duh, radio dosen't exist on init
 
-	//PDA
-	pda = new(src)
-	spawn(5)
-		pda.ownjob = "pAI Messenger"
-		pda.owner = text("[]", src)
-		pda.name = pda.owner + " (" + pda.ownjob + ")"
+	make_laws()
 
-	possible_chassis = typelist(NAMEOF(src, possible_chassis), list("cat" = TRUE, "mouse" = TRUE, "monkey" = TRUE, "corgi" = FALSE,
-									"fox" = FALSE, "repairbot" = TRUE, "rabbit" = TRUE, "borgi" = FALSE ,
-									"parrot" = FALSE, "bear" = FALSE , "mushroom" = FALSE, "crow" = FALSE ,
-									"fairy" = FALSE , "spiderbot" = FALSE))		//assoc value is whether it can be picked up.
+	possible_chassis = typelist(NAMEOF(src, possible_chassis), 
+		list("cat" = TRUE, "mouse" = TRUE, "monkey" = TRUE, "corgi" = FALSE,
+			"fox" = FALSE, "repairbot" = TRUE, "rabbit" = TRUE, "borgi" = FALSE ,
+			"parrot" = FALSE, "bear" = FALSE , "mushroom" = FALSE, "crow" = FALSE ,
+			"fairy" = FALSE , "spiderbot" = FALSE))		//assoc value is whether it can be picked up.
 	dynamic_chassis_icons = typelist(NAMEOF(src, dynamic_chassis_icons), initialize_dynamic_chassis_icons())
 	chassis_pixel_offsets_x = typelist(NAMEOF(src, chassis_pixel_offsets_x), default_chassis_pixel_offsets_x())
 
 	. = ..()
+	// START_PROCESSING(SSfastprocess, src) //fire-off when the things has to init is done
 
+	//PDA
+	pda.ownjob = "pAI Messenger"
+	pda.owner = "[src]"
+	pda.name = "[pda.owner] ([pda.ownjob])"
+
+	//Action buttons
 	var/datum/action/innate/pai/software/SW = new
-	var/datum/action/innate/pai/shell/AS = new /datum/action/innate/pai/shell
-	var/datum/action/innate/pai/chassis/AC = new /datum/action/innate/pai/chassis
-	var/datum/action/innate/pai/rest/AR = new /datum/action/innate/pai/rest
-	var/datum/action/innate/pai/light/AL = new /datum/action/innate/pai/light
-
-	var/datum/action/language_menu/ALM = new
 	SW.Grant(src)
+	var/datum/action/innate/pai/shell/AS = new
 	AS.Grant(src)
+	var/datum/action/innate/pai/chassis/AC = new
 	AC.Grant(src)
+	var/datum/action/innate/pai/rest/AR = new
 	AR.Grant(src)
+	var/datum/action/innate/pai/light/AL = new
 	AL.Grant(src)
+	var/datum/action/language_menu/ALM = new
 	ALM.Grant(src)
-	emitter_next_use = world.time + 10 SECONDS
+
+	emitter_next_use = world.time + (10 SECONDS)
+	GLOB.pai_list |= src
 
 /mob/living/silicon/pai/ComponentInitialize()
 	. = ..()
 	if(possible_chassis[chassis])
 		AddElement(/datum/element/mob_holder, chassis, 'icons/mob/pai_item_head.dmi', 'icons/mob/pai_item_rh.dmi', 'icons/mob/pai_item_lh.dmi', ITEM_SLOT_HEAD)
 
-/mob/living/silicon/pai/BiologicalLife(seconds, times_fired)
-	if(!(. = ..()))
-		return
-	if(hacking)
-		process_hack()
-
 /mob/living/silicon/pai/proc/process_hack()
-
-	if(cable && cable.machine && istype(cable.machine, /obj/machinery/door) && cable.machine == hackdoor && get_dist(src, hackdoor) <= 1)
+	if(cable?.machine && istype(cable.machine, /obj/machinery/door) && cable.machine == hackdoor && get_dist(src, hackdoor) <= 1)
 		hackprogress = clamp(hackprogress + 4, 0, 100)
 	else
 		temp = "Door Jack: Connection to airlock has been lost. Hack aborted."
@@ -188,12 +227,12 @@
 	..()
 	if(statpanel("Status"))
 		if(!stat)
-			stat(null, text("Emitter Integrity: [emitterhealth * (100/emittermaxhealth)]"))
+			stat(null, "Emitter Integrity: [emitterhealth * (100/emittermaxhealth)]")
 		else
-			stat(null, text("Systems nonfunctional"))
+			stat(null, "Systems nonfunctional")
 
 /mob/living/silicon/pai/restrained(ignore_grab)
-	. = FALSE
+	return FALSE
 
 // See software.dm for Topic()
 
@@ -203,77 +242,32 @@
 		return FALSE
 	return TRUE
 
-/mob/proc/makePAI(delold)
-	var/obj/item/paicard/card = new /obj/item/paicard(get_turf(src))
-	var/mob/living/silicon/pai/pai = new /mob/living/silicon/pai(card)
+/// turn the current mob into a pAI
+/mob/proc/makePAI(delold = FALSE)
+	if(istype(src, /mob/living/silicon/pai))
+		return FALSE
+	var/obj/item/paicard/card = new(get_turf(src))
+	var/mob/living/silicon/pai/pai = new(card)
 	transfer_ckey(pai)
 	pai.name = name
 	card.setPersonality(pai)
 	if(delold)
 		qdel(src)
+	return TRUE
 
-/datum/action/innate/pai
-	name = "PAI Action"
-	icon_icon = 'icons/mob/actions/actions_silicon.dmi'
-	var/mob/living/silicon/pai/P
-
-/datum/action/innate/pai/Trigger()
-	if(!ispAI(owner))
-		return 0
-	P = owner
-
-/datum/action/innate/pai/software
-	name = "Software Interface"
-	button_icon_state = "pai"
-	background_icon_state = "bg_tech"
-
-/datum/action/innate/pai/software/Trigger()
+/obj/item/paicard/attackby(obj/item/W, mob/user, params)
 	..()
-	P.paiInterface()
+	user.set_machine(src)
+	if(!W.tool_behaviour == TOOL_SCREWDRIVER && !istype(W, /obj/item/encryptionkey))
+		return FALSE
 
-/datum/action/innate/pai/shell
-	name = "Toggle Holoform"
-	button_icon_state = "pai_holoform"
-	background_icon_state = "bg_tech"
-
-/datum/action/innate/pai/shell/Trigger()
-	..()
-	if(P.holoform)
-		P.fold_in(FALSE)
+	if(pai?.encryptmod)
+		pai.radio.attackby(W, user, params)
 	else
-		P.fold_out()
+		to_chat(user, "Encryption Key ports not configured.")
 
-/datum/action/innate/pai/chassis
-	name = "Holochassis Appearance Composite"
-	button_icon_state = "pai_chassis"
-	background_icon_state = "bg_tech"
-
-/datum/action/innate/pai/chassis/Trigger()
-	..()
-	P.choose_chassis()
-
-/datum/action/innate/pai/rest
-	name = "Rest"
-	button_icon_state = "pai_rest"
-	background_icon_state = "bg_tech"
-
-/datum/action/innate/pai/rest/Trigger()
-	..()
-	P.lay_down()
-
-/datum/action/innate/pai/light
-	name = "Toggle Integrated Lights"
-	icon_icon = 'icons/mob/actions/actions_spells.dmi'
-	button_icon_state = "emp"
-	background_icon_state = "bg_tech"
-
-/datum/action/innate/pai/light/Trigger()
-	..()
-	P.toggle_integrated_light()
-
-/mob/living/silicon/pai/Process_Spacemove(movement_dir = 0)
-	. = ..()
-	if(!.)
+/mob/living/silicon/pai/Process_Spacemove(movement_dir = NONE)
+	if(!(. = ..()))
 		add_movespeed_modifier(/datum/movespeed_modifier/pai_spacewalk)
 		return TRUE
 	remove_movespeed_modifier(/datum/movespeed_modifier/pai_spacewalk)
@@ -296,6 +290,9 @@
 	if(!(. = ..()))
 		return
 	silent = max(silent - 1, 0)
+	emitterhealth = clamp((emitterhealth + emitterregen), -50, emittermaxhealth) //biolife ticks the same as ssfastproccess right?
+	if(hacking)
+		process_hack()
 
 /mob/living/silicon/pai/updatehealth()
 	if(status_flags & GODMODE)
@@ -303,23 +300,10 @@
 	health = maxHealth - getBruteLoss() - getFireLoss()
 	update_stat()
 
-/mob/living/silicon/pai/process()
-	emitterhealth = clamp((emitterhealth + emitterregen), -50, emittermaxhealth)
-
-/obj/item/paicard/attackby(obj/item/W, mob/user, params)
-	..()
-	user.set_machine(src)
-	var/encryption_key_stuff = W.tool_behaviour == TOOL_SCREWDRIVER || istype(W, /obj/item/encryptionkey)
-	if(!encryption_key_stuff)
-		return
-	if(pai?.encryptmod)
-		pai.radio.attackby(W, user, params)
-	else
-		to_chat(user, "Encryption Key ports not configured.")
-
 /mob/living/silicon/pai/proc/short_radio()
 	if(radio_short_timerid)
 		deltimer(radio_short_timerid)
+
 	radio_short = TRUE
 	to_chat(src, "<span class='danger'>Your radio shorts out!</span>")
 	radio_short_timerid = addtimer(CALLBACK(src, .proc/unshort_radio), radio_short_cooldown, flags = TIMER_STOPPABLE)
@@ -445,3 +429,63 @@
 	.["Cyborg - Security (dog - valesci)"] = -16
 	//Misc
 	.["Cyborg - Misc (dog - blade)"] = -16
+
+
+/datum/action/innate/pai
+	name = "PAI Action"
+	icon_icon = 'icons/mob/actions/actions_silicon.dmi'
+	var/mob/living/silicon/pai/P
+
+/datum/action/innate/pai/Trigger()
+	if(!ispAI(owner))
+		return FALSE
+	P = owner
+
+/datum/action/innate/pai/software
+	name = "Software Interface"
+	button_icon_state = "pai"
+	background_icon_state = "bg_tech"
+
+/datum/action/innate/pai/software/Trigger()
+	..()
+	P.paiInterface()
+
+/datum/action/innate/pai/shell
+	name = "Toggle Holoform"
+	button_icon_state = "pai_holoform"
+	background_icon_state = "bg_tech"
+
+/datum/action/innate/pai/shell/Trigger()
+	..()
+	if(P.holoform)
+		P.fold_in(FALSE)
+	else
+		P.fold_out()
+
+/datum/action/innate/pai/chassis
+	name = "Holochassis Appearance Composite"
+	button_icon_state = "pai_chassis"
+	background_icon_state = "bg_tech"
+
+/datum/action/innate/pai/chassis/Trigger()
+	..()
+	P.choose_chassis()
+
+/datum/action/innate/pai/rest
+	name = "Rest"
+	button_icon_state = "pai_rest"
+	background_icon_state = "bg_tech"
+
+/datum/action/innate/pai/rest/Trigger()
+	..()
+	P.lay_down()
+
+/datum/action/innate/pai/light
+	name = "Toggle Integrated Lights"
+	icon_icon = 'icons/mob/actions/actions_spells.dmi'
+	button_icon_state = "emp"
+	background_icon_state = "bg_tech"
+
+/datum/action/innate/pai/light/Trigger()
+	..()
+	P.toggle_integrated_light()
